requirements.txt
python-telegram-bot
pillow
numpy